Fix IPv6 regular expression and console error.
The previous regex failed for a couple of valid IPv6 address notations. This commit updates the regex to work for (almost) all valid IPv6 address notations. Also it adds an additional check for the result of the exec call to prevent console error messages in cases where the regex fails.